frontend+backend+model/requirements.txt
# Core NLP and utilities
spacy>=3.7.2
pydantic>=2.7
python-dotenv>=1.0

# LangChain + Google Gemini (v1)
langchain>=0.3.7
langchain-community>=0.3.7
langchain-google-genai>=2.0.2
google-generativeai>=0.8.3

# Vector store: PostgreSQL + pgvector
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.10
pgvector>=0.2.5

# Document loaders
pypdf>=5.0.0
python-docx>=1.1.2
unstructured>=0.15.9
unstructured[pdf]>=0.15.9

# Web framework (example API)
fastapi>=0.114
uvicorn>=0.31

# Optional utilities
tenacity>=8.3